Replace all u64/i64 for length, index or offsets into usize/isize
On platforms that are 32-bits (or 128?) this should make array and
iterations faster in general. On platforms that are 64-bits, this
will have no effect.

+++ core/engine/src/builtins/array/tests.rs
@@ -874,7 +874,7 @@ fn array_spread_non_iterable() {
 #[test]
 fn get_relative_start() {
     #[track_caller]
-    fn assert(context: &mut Context, arg: Option<&JsValue>, len: u64, expected: u64) {
+    fn assert(context: &mut Context, arg: Option<&JsValue>, len: usize, expected: usize) {
         let arg = arg.unwrap_or(&JsValue::Undefined);
         assert_eq!(
             Array::get_relative_start(context, arg, len).unwrap(),
@@ -901,7 +901,7 @@ fn get_relative_start() {
 #[test]
 fn get_relative_end() {
     #[track_caller]
-    fn assert(context: &mut Context, arg: Option<&JsValue>, len: u64, expected: u64) {
+    fn assert(context: &mut Context, arg: Option<&JsValue>, len: usize, expected: usize) {
         let arg = arg.unwrap_or(&JsValue::Undefined);
         assert_eq!(
             Array::get_relative_end(context, arg, len).unwrap(),
