+++ core/engine/src/object/builtins/jsarraybuffer.rs
@@ -60,7 +60,7 @@ impl JsArrayBuffer {
                 .array_buffer()
                 .constructor()
                 .into(),
-            byte_length as u64,
+            byte_length,
             None,
             context,
         )?;
